job-portal-backend_requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.0
pypdf2==3.0.1
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
gunicorn==21.2.0